End of section 18

+++ client/src/app/messages/messages.component.ts
@@ -2,6 +2,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
 import { Pagination } from '../_models/pagination';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component({
@@ -16,7 +17,7 @@ export class MessagesComponent implements OnInit {
   pageNumber=1;
   pageSize=5;
   loading = false;
-  constructor(private messageService: MessageService) { }
+  constructor(private messageService: MessageService, private confirmService: ConfirmService) { }
 
   ngOnInit(): void {
     this.loading=true;
@@ -33,10 +34,18 @@ export class MessagesComponent implements OnInit {
     })
   }
   deleteMessage(id: number) {
-    this.messageService.deleteMessage(id).subscribe(() => {
+    this.confirmService.confirm('Confirm delete message', 'This cannot be undone')
+    .subscribe(result=>{
+        if(result)
+        {
+
+          this.messageService.deleteMessage(id).subscribe(() => {
       //1-how many messages we want to delete
-      this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
-    })
+          this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
+          }) 
+       }
+      }
+    )
   }
 
   pageChanged(event:any){
